SERVER-7322 Fix Linux compile
Some versions of g++ issue a warning on copying a double to an int:

src/mongo/db/geo/2d.cpp:79: warning: converting to 'int' from 'double'

Other g++ versions do not complain.  Use static_cast<int> to prevent warning.

+++ src/mongo/db/geo/2d.cpp
@@ -76,7 +76,7 @@ namespace mongo {
                 } else {
                     int order = 1;
                     if (e.isNumber()) {
-                        order = e.Number();
+                        order = static_cast<int>(e.Number());
                     }
                     _other.push_back(make_pair(e.fieldName(), order));
                 }
